backend/requirements.txt
alembic==1.13.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
bcrypt==4.0.1
certifi==2026.4.22
cffi==2.0.0
click==8.3.3
colorama==0.4.6
cryptography==47.0.0
dnspython==2.8.0
ecdsa==0.19.2
email-validator==2.3.0
fastapi==0.111.0
fastapi-cli==0.0.24
greenlet==3.4.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.27.0
idna==3.13
iniconfig==2.3.0
Jinja2==3.1.6
joblib==1.4.2
Mako==1.3.11
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
numpy==1.26.4
orjson==3.11.8
packaging==26.2
passlib==1.7.4
pluggy==1.6.0
psycopg2-binary==2.9.9
pyasn1==0.6.3
pycparser==3.0
pydantic==2.7.1
pydantic-settings==2.2.1
pydantic_core==2.18.2
Pygments==2.20.0
pytest==8.2.0
pytest-asyncio==0.23.6
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.3
rich==15.0.0
rich-toolkit==0.19.7
rsa==4.9.1
ruff==0.4.4
scikit-learn==1.4.2
scipy==1.17.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.30
starlette==0.37.2
threadpoolctl==3.6.0
typer==0.25.0
typing-inspection==0.4.2
typing_extensions==4.15.0
ujson==5.12.0
uvicorn==0.29.0
watchfiles==1.1.1
websockets==16.0
